pyproject.toml
[project]
name = "sirqle"
version = "0.2.5"
authors = [{ name = "Pythia Dev Team", email = "dev@pythia.social" }]
description = "SurrealDB Query interface"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["python-dotenv", "surrealdb>=0.3.0"]
license = { file = "LICENSE" }
[project.urls]
Source = "https://github.com/PythiaSocialTech/sirqle"

[tool.setuptools.packages.find]
where = ["src"]
include = ["sirqle*"]

[build-system]
requires = ["setuptools>=61.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
  "pre-commit",
  "pytest",
  "pytest-asyncio",
  "pytest-dependency",
  "commitizen",
  "flit",
]
docs = [
  "mkdocs",
  "mkdocstrings",
  'mkdocs-material',
  "mkdocs-literate-nav",
  "mkdocstrings[python]>=0.18",
  "mkdocs-gen-files",
  "mkdocs-literate-nav",
]

[tool.black]
line-length = 88

[tool.pycln]
all = true

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.2.5"
version_files = ["pyproject.toml:version"]
tag_format = "v$version"
bump_message = "bump: $current_version → $new_version [skip ci]"


[tool.pytest.ini_options]
testpaths = 'tests'
asyncio_mode = 'strict'
